Deleted cmake-.. directory to see consequences
Removed the cmake-build-debug directory because it was unnecessary for this project to run. This is an exercise to see what happens and for me to learn the best practices. I do not know if it is the correct action, but I can use the src file without any of the files present in cmake, except the "chatlog.txt"